+++ src/views/layout.tsx
@@ -9,6 +9,7 @@ import {
   HTMX_4XX_SWAP_SCRIPT,
   FORM_PERSIST_SCRIPT,
   EMAIL_VALIDATE_SCRIPT,
+  PROOF_UPLOAD_SCRIPT,
 } from "./scripts";
 
 // Meta carries page-level SEO/social metadata into the Layout.
@@ -47,6 +48,7 @@ const GlobalScripts: FC = () => (
     <script dangerouslySetInnerHTML={{ __html: HTMX_4XX_SWAP_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: FORM_PERSIST_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: EMAIL_VALIDATE_SCRIPT }} />
+    <script dangerouslySetInnerHTML={{ __html: PROOF_UPLOAD_SCRIPT }} />
   </>
 );